Overloaded notation for tactic sequencing.
Note that the associativity of this is different from the monadic ;;.

Note that contrary the earlier simicolon canonical structure, I made the
types of seq more restrictive, so that we get better errors, and so that
we can bind the arguments to the tactic scope. I tried doing this with a
canonical structure, but failed, so I used a type class.

s