Implement '<[decapp a b with x y z]>`.
This makes the `decapp` notation as powerful as the primitive by
allowing it to specify arguments to `b`. This is necessary in case of
depenent types where `b` only has `a`'s type if it is applied to `x y z`
first. The continuation does not take `x y z` as arguments - it only
takes the remaining ones.