Refactor Base.v into smaller files.
The resulting files mostly live in theories/intf/ – a place for files
that are needed both by the OCaml and the Coq side of Mtac.
theories/Pattern.v is the only exception here, as it is not actually
used in the OCaml code.

TODO: theories/intf/M.v could be split up further to extract `mmatch'`
and other derived functions.