Fixes #82.
This commit both simplifies and strengthens `M.decompose_app'`. We no
longer need partial telescopes. Instead, we take a candidate head
constant applied to some arguments. When no arguments are given we
simply compare the candidate with the head of the discriminant. When
the candidate is applied to arguments we unify them with the arguments
of the discriminant. Only when they match do we call the continuation.

Additionally, the return type `B` now depends on the discriminant. In
the continuation, we now demand `t (B (candidate arg1 … argN))`. This
way we can expose the fact that the match succeeded at the type level.