Generalize `mmatch` branches to decomp. prim's.
This introcudes three new kinds of `mmatch` branches:
1. `[!APP] f a .. c $[n|u|m|c] x .. z =n> b`, which compares the discriminant
with (f a' .. c' x' .. z'), unifying the initial arguments `a' .. c` with
`a .. c` according to the unification strategy indicated by the postfix to
`$`. The remaining arguments `x' .. z'` are bound variables in the body `b`.
2. `[!Type] forall _ : X, P =n> b` which compares the discriminants
against `forall x : X, P x` where `P : X -> Type`. Both `X` and `P` are
bound variables in the body `b`.
3. `[!Prop] forall _ : X, P =n> b`; a `Prop` version of the previous
branch.

All new branches use `=n>` to indicate that they reduce neither the
discriminant nor the candidate. the `[!APP]` pattern is the only pattern
that is able to perform unification at all, and even then it only does
so on the initial arguments, never on the full term.