Add `mrun_static` Ltac which avoids dyn. checks.
`mrun_static` takes a constant (not arbitrary constrs) and statically
ensures that it is either an `M …` or a `gtactic`. In the latter case,
no more additional runtime checks are performed when executing the
tactic. For `M T`, we still need to dynamically ensure that `T` matches
the current goal.