Fix `mrun_static`'s handling of universes.
The type stored by `mrun_static` was meaningless because it was only
valid in the original `sigma` used for typechecking, but not in the
`sigma` used when executing the tactic. The new schemes re-instantiates
the universes correctly. This may also reduce the number of universes in
`sigma` during execution: PMP conjectures that unifying the type of the
tactic with the goal type may have introduced the required
universes — which would then be introduced *again* by `fresh_global.