Prevent exceptions from leaking `nu` vars.
This was previously only enforced by calling `Typing.type_of` either at
the surrounding `mtry` or when the exception reaches the user. This
could lead to `nu` vars being leaked by an exception and then used
outside of the `nu`.

This commits add a test case to enforce this new behavior. The existing
test case meant to enforce this,
`tests/exceptions.nu_not_closed_raise_not_closed`, succeeded only
coincidentally because of the `Typing.type_of` call for the surrounding
`mtry`.