Define MTele functions in terms of `ArgsOf`.
This adds a lot of (un)currying functions and rewrites existing
functions to use these new functions.

NOTE 1: My initial instinct of avoiding explicit packing of arguments
via `ArgsOf` was misguided. `MTele_In`, which is the result of this
instinct, is still used in various places but should be avoided in the
future. The reason is that even with the accessor style introduced in
commit ea120f5d7ff2f5df53efe7fed69efb0f4e968f42 we still cannot express
all the relations between telescoped types and their values. In
particular, we cannot retroactively teach the accessor definition how to
access new telescope constructs that are defined using `MTele_In`. For
this reason, it is now recommended to use uncurrying whenever possible.

NOTE 2: Pattern.v contains changes that are necessary to convince Coq that
everything still typechecks. These changes, however, are not
theoretically necessary. The necessity of the type annotation on
`{m:MTele}` in particular is extremely baffling.