Unify `mmatch` syntax.
This commit introduces canonical structure infrastructure and instances
for dependent and independent `mmatch` syntax that can be shared between
`M` and `gtactic`. Apart from the general satisfaction of having just
one set of notations for `mmatch`, this also removes the syntactic
restriction on the return type of `mmatch` terms. In particular, this
allows us to write annotations like `return ttac ..`.

The infrastructure is split into two structures: `InDepMatcher` and
`Matcher`. `InDepMatcher` forces Coq to discard any possible dependency
on the term being inspected and thus helps the typechecker. `Matcher` is
used for dependent `mmatch` terms. It uses a helper structure,
`Predicate` to allow Coq to find the right monad (`M`, `gtactic`) based
on the return type as well as the type of branches.