`[#]` patterns: support primitive projections.
This change adds transparent support for primitive projections in `[#]`
patterns and, by virtue of it being the underlying primitive, to
`is_head`.

The implementation tries it best to be efficient and also follow the
users wishes. It will enforce unifiability of parameters if they are
part of the pattern (i.e. if the pattern itself is not a primitive
projection) and otherwise only unify the record value.

The code contains rather detailed comments since these things tend to be
rather tricky. Test cases have been added as far as possible but there
is one case that seems impossible to trigger (primitive pattern,
non-primitive discriminee) because Coq unfolds the compatibility
constants for projections rather eagerly.