Faster Coq string <-> OCaml strings transformation.
`to_coq` was creating N many substrings out of the original string and
also constructed the Coq string in a non-tail-recursive way, blowing up
the stack even for reasonably small strings.

`from_coq` now uses a string buffer, which may or may not be the same as
the original code but this way I know for sure what is happening.