Clean up mmatch, remove `.. => [H] ..` patterns.
This patch simplifies the types `branch` and `pattern` by removing the
argument `y : A`, i.e. the discriminee, from their signature. This
argument was used in two places: (1) to type the body of `pany`, and 
(2) to type the equality in `.. => [H] ..` patterns, which this patch 
removes completely.

Regarding (1): Since `pattern` no longer has access to the discriminee,
this patch changes the type of `pany` to 
`(forall x : A, B x) -> pattern A B` and introduces a new pattern notation 
written `_ as binder => ..` where the `_` is a fixed token like it was
in the old `_ => ..` notation. This syntax translates to 
`pany (fun x => ..)`. When using this notation, `binder` cannot be replaced 
by `_` since Coq uses `binder` to type the branch.

Regarding (2): We make little use of the `[H]` and all occurrences
can be replaced by uses of `mtmmatch`, which is surprisingly usable at
this point. Switching to `mtmmatch` currently means that fewer patterns
are available since it has not been ported to the new `branch` and
`pattern` infrastructure. Therefore, some `_ => ..` branches have been
translated to `t =n> ..` where `t` is the discriminee of the match.